requirements.txt
pyannote.audio==3.3.0
torch==2.4.1
pydub==0.25.1
openai
yt_dlp